+++ packages/web/src/web/pageVisibility.ts
@@ -1,18 +1,22 @@
 import { ref, Ref } from "@vue/composition-api";
 
 let visibility: Ref<VisibilityState> | undefined = undefined;
-let hidden = ref(document.hidden);
+let hidden: Ref<boolean> | undefined = undefined;
 
 export function usePageVisibility() {
+  if (!hidden) {
+    hidden = ref(document.hidden);
+  }
   if (!visibility) {
     visibility = ref(document.visibilityState);
     document.addEventListener(
       "visibilitychange",
       () => {
         visibility!.value = document.visibilityState
-        hidden.value = document.hidden;
+        hidden!.value = document.hidden;
       },
       { passive: true }
+      // true
     );
   }
   return {