+++ packages/web/src/web/language.ts
@@ -1,18 +1,18 @@
 import { ref, Ref } from "@vue/composition-api";
 
 let language: Ref<String> | undefined = undefined;
-let preferred: Ref<readonly String[]> | undefined = undefined;
+let languages: Ref<readonly String[]> | undefined = undefined;
 
 export function useLanguage() {
   if (!language) {
     language = ref(navigator.language);
   }
 
-  if (!preferred) {
-    preferred = ref(navigator.languages);
+  if (!languages) {
+    languages = ref(navigator.languages);
     const change = () => {
       language!.value = navigator.language;
-      preferred!.value = navigator.languages;
+      languages!.value = navigator.languages;
     };
 
     window.addEventListener('languagechange', change, { passive: true });
@@ -20,6 +20,6 @@ export function useLanguage() {
 
   return {
     language,
-    preferred
+    languages
   };
 }