+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased]
+
+--- -->
+
+## 1.0.0-dev.6
+
+_2020-01-19_
 
 ---